+++ src/box/packager.py
@@ -63,8 +63,13 @@ class PackageApp:
 
     @property
     def builders(self) -> List:
-        """Return a list of supported builders and their commands."""
+        """Return a list of supported builders."""
         return list(self._builders.keys())
+
+    @property
+    def builders_and_custom(self) -> List:
+        """Return a list of supported builders and custom builder."""
+        return self.builders + ["custom"]
 
     @property
     def binary_name(self):
@@ -77,14 +82,31 @@ class PackageApp:
             self._config = PyProjectParser()
         return self._config
 
+    @property
+    def builder_command(self) -> List[str]:
+        """Get the command list to run for specific builder.
+
+        :param builder: Builder to run with.
+
+        :raises KeyError: Unknown builder.
+        """
+        builder = self.config.builder
+
+        if builder.lower().startswith("custom"):
+            cmd = builder.split("=", 1)[1].strip("'\"")
+            return cmd.split(" ")
+
+        try:
+            return self._builders[builder]
+        except KeyError as e:
+            raise KeyError(f"Unknown {builder=}") from e
+
     def build(self):
         """Build the project with PyApp."""
         builder = self.config.builder
         fmt.info(f"Building project with {builder}...")
-        try:
-            subprocess.run(self._builders[builder], **self.subp_kwargs)
-        except KeyError as e:
-            raise KeyError(f"Unknown {builder=}") from e
+
+        subprocess.run(self.builder_command, **self.subp_kwargs)
 
         fmt.success(f"Project built with {builder}.")